requirements.txt
ruff==0.6.7
openpyxl==3.1.5
pillow==10.4.0